EdViz-Core/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pdfplumber==0.10.3
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
slowapi==0.1.8
pydantic==2.6.1
sqlmodel>=0.0.8
asyncpg>=0.25.0
email-validator==2.1.0.post1
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
openai>=1.0.0 